correct date milliseconds fields

+++ composeApp/src/desktopMain/kotlin/io/haskins/staffmanagement/dao/models/EmployeeNotes.kt
@@ -6,7 +6,7 @@ import org.jetbrains.exposed.sql.Table
 object EmployeeNotes: Table() {
     val id: Column<Int> = integer("en_id").autoIncrement()
 
-    val date: Column<Int> = integer("date")
+    val date: Column<Long> = long("date")
 
     val employeeId: Column<Int> = (integer("employee_id") references Projects.id)
     val title: Column<String> = text("title")
