+++ composeApp/src/desktopMain/kotlin/io/haskins/staffmanagement/dao/models/ProjectNotes.kt
@@ -6,7 +6,7 @@ import org.jetbrains.exposed.sql.Table
 object ProjectNotes: Table() {
     val id: Column<Int> = integer("pn_id").autoIncrement()
 
-    val date: Column<Int> = integer("date")
+    val date: Column<Long> = long("date")
 
     val projectId: Column<Int> = (integer("project_id") references Projects.id)
     val title: Column<String> = text("title")
