+++ composeApp/src/desktopMain/kotlin/io/haskins/staffmanagement/dao/models/ProjectResources.kt
@@ -6,8 +6,8 @@ import org.jetbrains.exposed.sql.Table
 object ProjectResources: Table() {
     val id: Column<Int> = integer("pr_id").autoIncrement()
 
-    val start: Column<Int> = integer("start")
-    val end: Column<Int> = integer("end")
+    val start: Column<Long> = long("start")
+    val end: Column<Long> = long("end")
 
     val employeeId: Column<Int> = (integer("employee_id") references Employees.id)
     val projectId: Column<Int> = (integer("project_id") references Projects.id)
